llm-analytics/instructor-python/requirements.txt
posthog>=3.0.0
instructor>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
